webmidi/midioutput_send.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="UTF-8">
    <title>Web MIDI API Test: send()</title>
    <link rel="author" title="Ryoya KAWAI" href="mailto:ryoya.kawai@gmail.com">
    <link rel="help" href="http://webaudio.github.io/web-midi-api/#obtaining-access-to-midi-devices">
    <link rel="match" href="reference/REFERENCE-FILE-NAME">
    <meta name="assert" content="Success when MIDI messages send to device with send() and MIDI device plays C4(60) D4(62) E4(64) F4(65) G4(67) A4(69) B4(71) C5(72) in Piano voice.">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <style type="text/css">
    /* ADD STYLE BLOCK HERE (PREFERRABLE TO INLINE STYLES) */
    body { font-family:arial;}
    </style>
  </head>
  <body>
    <p>Success when MIDI messages send to device with send() and MIDI device plays C4(60) D4(62) E4(64) F4(65) G4(67) A4(69) B4(71) C5(72) in Piano voice.</p>
    <div id="warning" style="color:red;"></div>
    <div id="selectdevice"></div>
    <p>Selected Device: <span id="outputdevice" style="border:1px solid #afafaf; border-radius:4px; padding:2px 4px;"></span></p>
    <div id="log"></div>

    <!-- PAGE CONTENT -->
    <!--script type="text/javascript" src="webmidiapi_07_send.js"></script-->
    <script type="text/javascript">
    var MIDI;
    var timerId={ };
    var getPort={ };
    var outputs=[];

    var test_outputs = async_test(
        "send() - W3C Spec: Web MIDI API",
        {timeout: 5000} //msec
    );

    navigator.requestMIDIAccess().then( successCallback, errorCallback );
    function successCallback(access) {
        MIDI = access;
        test_outputs.step(function() {
            if(typeof MIDI.outputs==="function") {
                document.getElementById("warning").innerHTML= "This browser seems the implementation of the API, Web MIDI API, is followed old spec.";
                outputs=MIDI.outputs();
            } else {
                var outputIterator=MIDI.outputs.values();
                for(var o=outputIterator.next(); !o.done; o=outputIterator.next()) {
                    outputs.push(o.value);
                }
            }
            console.log("[OutputDevices] ", outputs.length);
            if(outputs.length==0) {
                document.getElementById("warning").innerHTML="Your Computer looks no <b><u>MIDI OUTPUT</u></b> devices are pluged-in.<br> Please plug-in at least one <b><u>MIDI OUTPUT</u></b> device before run this Test.";
                assert_false(true);
                disableButton()
                test_outputs.done();
            } else {

                var input = document.createElement("input");
                input.type="button"; input.id="start"; input.value="Run";
                var select = document.createElement("select");
                select.id="midiout";
                for(var i=0; i<outputs.length; i++) {
                    select.options[i]=new Option(outputs[i]["name"]+" ("+outputs[i]["id"]+")", i);
                }
                document.getElementById("selectdevice").appendChild(select);
                document.getElementById("selectdevice").appendChild(input);

                document.getElementById("start").addEventListener("click", function(){
                    var selIdx=document.getElementById("midiout").selectedIndex;
                    //disableButton()

                    var o = outputs[selIdx];
                    document.getElementById("outputdevice").innerHTML=outputs[selIdx]["name"]+" ("+outputs[selIdx]["id"]+")";
                    assert_true(outputs instanceof Object);

                    var count=0, j=0;
                    var msg=[0x3c, 0x3e, 0x40, 0x41, 0x43, 0x45, 0x47, 0x48];
                    var timerId = setInterval(function(){
                        j++;
                        var fst=0x80;
                        if(j%2==1) {
                            fst=0x90;
                        }
                        var snd=msg[count];
                        o.send( [ 0xc0, 0x01] );
                        o.send( [ fst, snd, 0x7f ] );
                        console.log(fst.toString(16), snd.toString(16), parseInt(0x7f).toString(16));
                        if(j%2==0) count++;
                        if(msg.length<=count) {
                            clearInterval(timerId);
                            assert_true(true);
                            disableButton();
                            test_outputs.done();
                        }}, 100); //default:70
                    });

                    // automatic
                    if(typeof outputs=="object" && outputs.length>0) {
                        var type="click";
                        var e=document.createEvent('MouseEvent');
                        var b=document.getElementById("start");
                        e.initEvent(type, true, true, window, 1, 0, 0, 0, 0, false, false, false, false, 0, null);
                        //b.dispatchEvent(e);
                    }
                    
                }
            });
       }
        function errorCallback(message) {
            test_outputs.step(function() {
                assert_true(false);
                disableButton()
                test_outputs.done();
            });
        }
        function disableButton() {
            document.getElementById("midiout").setAttribute("disabled", "disabled");
            document.getElementById("start").setAttribute("disabled", "disabled");
        }
        
    </script>

  </body>
</html>
